Move dict_to_querydict() and normalize_querydict() to utilities.querydict
