+++ netbox/extras/dashboard/widgets.py
@@ -16,8 +16,9 @@ from core.models import ObjectType
 from extras.choices import BookmarkOrderingChoices
 from netbox.choices import ButtonColorChoices
 from utilities.permissions import get_permission_for_model
+from utilities.querydict import dict_to_querydict
 from utilities.templatetags.builtins.filters import render_markdown
-from utilities.utils import content_type_identifier, content_type_name, dict_to_querydict
+from utilities.utils import content_type_identifier, content_type_name
 from utilities.views import get_viewname
 from .utils import register_widget
