+++ netbox/netbox/views/generic/object_views.py
@@ -18,7 +18,8 @@ from utilities.error_handlers import handle_protectederror
 from utilities.exceptions import AbortRequest, PermissionsViolation
 from utilities.forms import ConfirmationForm, restrict_form_fields
 from utilities.permissions import get_permission_for_model
-from utilities.utils import normalize_querydict, prepare_cloned_fields
+from utilities.querydict import normalize_querydict
+from utilities.utils import prepare_cloned_fields
 from utilities.views import GetReturnURLMixin, get_viewname
 from .base import BaseObjectView
 from .mixins import ActionsMixin, TableMixin
